@@ -2,8 +2,8 @@
   "name": "angular-rest",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^6.0.0-rc.0 || ^6.0.0",
-    "@angular/core": "^6.0.0-rc.0 || ^6.0.0"
+    "@angular/common": "^6.1.0",
+    "@angular/core": "^6.1.0"
   },
   "dependencies": {
     "@angular-redux/form": "^9.0.1",
@@ -14,5 +14,10 @@
     "moment-timezone": "^0.5.21",
     "redux": "^4.0.0",
     "redux-observable": "^1.0.0"
+  },
+  "devDependencies": {
+    "@commitlint/cli": "^7.0.0",
+    "@commitlint/config-conventional": "^7.0.1",
+    "tslint": "^5.11.0"
   }
 }
