@@ -1,21 +1,16 @@
-import { Injectable } from '@angular/core';
-import { RestEntity } from './domain/api/rest-entity';
-import { NgRedux } from '../../node_modules/@angular-redux/store';
-import { IAppState } from './stores/root.store';
-import { RestEntityManager } from './domain/api/rest-entity-manager';
-import { AbstractEntity } from './domain/entities/abstract.entity';
+import {Injectable} from '@angular/core';
+import {NgRedux} from '@angular-redux/store';
+import {IAppState} from './stores/root.store';
+import {EntityManager} from './domain/api/entity-manager';
+import {AbstractEntity} from './domain/entities/abstract.entity';
 
 @Injectable({
   providedIn: 'root'
 })
 export class AngularRestService {
-  protected path: string;
-  protected resource: RestEntity;
-
   constructor(private ngRedux: NgRedux<IAppState>) {}
 
-  getManager(entityName: any): RestEntityManager<AbstractEntity> {
-    console.log(entityName.name);
+    getManager(entityName: any): EntityManager<AbstractEntity> {
     return this.ngRedux.getState()[entityName.name].get('manager');
   }
 }
