@@ -1,12 +1,10 @@
 export class ApiResponse {
     status: number;
-    function: string;
-    page: string;
     msg: any;
     data: any;
     requestData: any;
 
-    estValide(): boolean {
+    isValid(): boolean {
         return this.status >= 200 && this.status < 300;
     }
 }
