@@ -1,22 +1,21 @@
-import { combineReducers } from 'redux';
-import { RestReducer } from './rest/rest.reducer';
-import { entities } from '../domain/rest-entities';
-import { RestEntity } from '../domain/api/rest-entity';
-import { RestEntityManager } from '../domain/api/rest-entity-manager';
+import {combineReducers, ReducersMapObject} from 'redux';
+import {RestReducer} from './rest/rest.reducer';
+import {entitiesList} from '../domain/entities.list';
+import {EntityManager} from '../domain/api/entity-manager';
+import {EntityDescriptor} from "../domain/descriptors";
 
 export interface IAppState {
     [key: string]: any;
 }
 
 export class RootReducer {
-  public static getReducer() {
+    public static getReducer() {
+        const reducers: ReducersMapObject = {};
 
-    const reducers = {};
+        entitiesList.forEach((re: EntityDescriptor) => {
+            reducers[re.class.name] = new RestReducer<typeof re.class>(new EntityManager<typeof re.class>(re.class)).createReducer();
+        });
 
-    entities.forEach((re: RestEntity) => {
-        reducers[re.class.name] = new RestReducer<typeof re.class>(new RestEntityManager<typeof re.class>(re.class)).createReducer();
-    });
-
-    return combineReducers(reducers);
-  }
+        return combineReducers(reducers);
+    }
 }
